requirements.txt
transitions==0.6.9
paho-mqtt==1.4.0
gpiozero==1.5.0
python-dotenv==0.10.3
Si7021==0.1.1
smbus2==0.2.3
inky==0.0.5
Adafruit_DHT==1.4.0
flask
flask_restful